[wip] eden.cpp cleanup
