+++ eden/Eden.cpp
@@ -19,13 +19,6 @@ Parallel simulation engine for ODE-based models
 
 //standard libs
 #include <cmath>
-#include <chrono>
-
-//MPI and GPU options
-#include "mpi_setup.h"
-#ifdef USE_GPU
-    #include "GPU_helpers.h"
-#endif
 
 #if defined (__linux__) || defined(__APPLE__)
 #include <dlfcn.h> // for dynamic loading
@@ -40,57 +33,55 @@ Parallel simulation engine for ODE-based models
 #include "Common.h"
 #include "NeuroML.h"
 #include "MMMallocator.h"
+#include "StringHelpers.h"
+
+//clean
+#include "GPU_helpers.h"
+#include "Timer.h"
 
 // mess to clean up
 #include "IterationCallback.h"
-#include "AppendToVector.h"
-#include "string_helpers.h"
 #include "FixedWidthNumberPrinter.h"
 #include "GeomHelp_Base.h"
-#include "RawTables.h"
-#include "TableEntry.h"
 #include "TypePun.h"
+
 #include "EngineConfig.h"
 #include "SimulatorConfig.h"
 #include "GenerateModel.h"
-#include "StateBuffers.h"
-#include "MpiBuffers.h"
+#include "Mpi_helpers.h"
 #include "TrajectoryLogger.h"
-#include "Timer.h"
 #include "backends.h"
 #include "parse_command_line_args.h"
 #include "print_eden_cli_header.h"
 
-void setup_gpu(EngineConfig &engine_config){
-    if(engine_config.backend == backend_kind_gpu){
-        printf("GPU backend selected, checking if it's available");
-    }
-    #ifdef USE_GPU
-    printf("hello from CPU compiled code\n");
-    test();
-    #else
-    printf("No GPU support");
-    #endif
-}
-
 int main(int argc, char **argv){
-    setvbuf(stdout, NULL, _IONBF, 0); // first of all, set stdout,stderr to Unbuffered, for live output
-    setvbuf(stderr, NULL, _IONBF, 0); // this action must happen before any output is written !
-    print_eden_cli_header();          // To print the CLI header
-    RunMetaData metadata;             // To keep track of non backend specific meta data
-    SimulatorConfig config;           // Configuration of the simulator
-    Model model;                      // Keeping track of the model
-    EngineConfig engine_config;       // Configuration of the Engine
-    AbstractBackend *backend = nullptr;
 
-// Check the command line input with options
+//-----> Starting the simulator
+    {
+        setvbuf(stdout, NULL, _IONBF, 0); // first of all, set stdout,stderr to Unbuffered, for live output
+        setvbuf(stderr, NULL, _IONBF, 0); // this action must happen before any output is written !
+        print_eden_cli_header();          // To print the CLI header
+    }
+
+//-----> declaration of all used variables
+    RunMetaData metadata;             // Struct for keeping track of non backend specific meta data
+    SimulatorConfig config;           // Struct for Configuration of the simulator
+    Model model;                      // Struct for Keeping track of the model
+    EngineConfig engine_config;       // Struct for Configuration of the Engine
+
+    AbstractBackend *backend = nullptr;             // Class to handle all backend calls
+    TrajectoryLogger *trajectory_logger = nullptr;  // Class to handle all output generation
+    MpiBuffers *mpi_buffers = nullptr;              // Class to handle all MPI communication
+
+//-----> Find and check the specific engine_config
+    setup_mpi(argc, argv);              //check if everything works fine
+    setup_gpu(engine_config);           //same for gpu
+
+//-----> Check the command line input with options
     parse_command_line_args(argc, argv, config, model, metadata.config_time_sec);
 
-// Find and check the specific engine_config
-    setup_mpi(argc, argv);
-    setup_gpu(engine_config);
-
-// Init the backend
+//-----> Init the backend
+    printf("Initializing backend...\n");
     {
         if (engine_config.backend == backend_kind_cpu) {
             backend = new CpuBackend();
@@ -100,34 +91,29 @@ int main(int argc, char **argv){
         }
     }
 
-
 //----> Initialize the backend
     printf("Initializing model...\n");
-    Timer init_timer;
-    if (!GenerateModel(model, config, engine_config, backend->tabs)) {
-        printf("NeuroML model could not be created\n");
-        exit(1);
+    {
+        Timer init_timer;
+        if (!GenerateModel(model, config, engine_config, backend->tabs)) {
+            printf("NeuroML model could not be created\n");
+            exit(1);
+        }
+        trajectory_logger = new TrajectoryLogger(engine_config); //To log results
+        printf("Allocating state buffers...\n");
+        backend->init();
+
+        //just some timer functions to time this meta data --> encorperate this into meta data class
+        metadata.init_time_sec = init_timer.delta();
+
+        //debug mem layout
+        if(config.dump_raw_layout) backend->state->dump_raw_layout(backend->tabs);
+
+        //Initialize MPI
+         mpi_buffers = new MpiBuffers(engine_config);
     }
 
-    TrajectoryLogger trajectory_logger(engine_config); //To log results
-
-    printf("Allocating state buffers...\n");
-    backend->init();
-
-    //just some timer functions to time this meta data --> encorperate this into meta data class
-    metadata.init_time_sec = init_timer.delta();
-//<-----
-
-
-//to backend.
-    if(config.dump_raw_layout) backend->state->dump_raw_layout(backend->tabs);
-
-
-// call this MPI communicator
-    MpiBuffers mpi_buffers(engine_config);
-
-//---->  Simulations loop
-
+//----> Simulations loop
     printf("Starting simulation loop...\n");
     {
         Timer run_timer;
@@ -139,14 +125,14 @@ int main(int argc, char **argv){
             bool initializing = step <= 0;
 
             //init mpi communication --> empty call if no mpi compilation
-            mpi_buffers.init_communicate(engine_config, backend->state, config); // need to copy between backend & state when using mpi
+            mpi_buffers->init_communicate(engine_config, backend->state, config); // need to copy between backend & state when using mpi
 
             //execute the actual work items
             backend->execute_work_items(engine_config, config, (int)step, time);
 
             //dont check on initializing check on step < 0
             if (!initializing) {
-                trajectory_logger.write_output_logs(engine_config, time,
+                trajectory_logger->write_output_logs(engine_config, time,
                                                     backend->global_state_now(), /* needed on mpi???: */backend->global_tables_stateNow_f32());
             }
 
@@ -154,7 +140,7 @@ int main(int argc, char **argv){
             backend->dump_iteration(config, initializing, time, step);
 
             //waith for all the MPI communication to be done.
-            mpi_buffers.finish_communicate();
+            mpi_buffers->finish_communicate();
 
             // check on step
             if (!initializing) time += engine_config.dt;
@@ -165,5 +151,11 @@ int main(int argc, char **argv){
         metadata.run_time_sec = run_timer.delta();
     }
 
+//----> Print meta overeview
     metadata.print();
+
+//-----> Terminating program
+    delete backend;
+    delete trajectory_logger;
+    delete mpi_buffers;
 }
