+++ eden/EngineConfig.h
@@ -1,11 +1,14 @@
+#ifndef ENGINECONFIG_H
+#define ENGINECONFIG_H
+
+#include "TableEntry.h"
+
 extern "C" {
 
 typedef uint32_t backend_kind;
 #define backend_kind_nil 0
 #define backend_kind_cpu 1
 #define backend_kind_gpu 2
-
-//todo clean this up
 
 // and more information that is needed for the engine
 struct EngineConfig{
@@ -76,3 +79,5 @@ struct EngineConfig{
 	
 };
 }
+
+#endif
