+++ eden/TypePun.h
@@ -1,3 +1,4 @@
+
 // smuggle int32 into f32 position
 // dodgy, but that's life
 union TypePun_I32F32{
